refactor: file structure
refactored the file structure, moved all constants
to a separate folder. Made sure all functions that weren't being tested
were not being run in our test file through imports.
