+++ .github/workflows/ci.yml
@@ -28,4 +28,4 @@ jobs:
       run: npm ci
 
     - name: Test
-      run:  npm tes
+      run:  npm test
