+++ src/constants/harvestInfo.js
@@ -1,0 +1,9 @@
+const HARVEST_TOKEN = '2719680.pt.8S1Yry48sFx_5qPZ4_a650DfG2Z4-_mwykp-uQTpWTVDttVS0CWxYah1wM_-mOdLNOdUpJ9QbdZFmap1UjgI_Q'; 
+const HARVEST_ACCOUNT_ID = 266954;
+const HARVEST_USER_ID = 3855900
+
+module.exports = {
+  HARVEST_TOKEN,
+  HARVEST_ACCOUNT_ID,
+  HARVEST_USER_ID
+};
