+++ src/constants/projectDictionaryAndOrder.js
@@ -1,0 +1,28 @@
+const PROJECT_DICTIONARY = {
+  DIRECTED_LEARNING: {
+    name: 'Directed Learning',
+    projectId: 3073688,
+    taskId: 1819355 
+  },
+  PAIRING: {
+    name: 'Pairing',
+    projectId: 3073697,
+    taskId: 1629017
+  },
+  GENERAL: {
+    name: 'General Time',
+    projectId: 3073665,
+    taskId: 1819358 
+  }
+}
+
+const ORDER = [
+  PROJECT_DICTIONARY.DIRECTED_LEARNING,
+  PROJECT_DICTIONARY.GENERAL,
+  PROJECT_DICTIONARY.PAIRING
+]
+
+module.exports = {
+  PROJECT_DICTIONARY,
+  ORDER
+};
