+++ src/js/displayHours.js
@@ -4,7 +4,7 @@ const { translateData } = require("./translateData");
 const displayProjectHours = async (project) => {
   
   const timeEntries = await getTimeEntries(project);
-  const translatedData = await translateData(timeEntries, project);
+  const translatedData = translateData(timeEntries, project);
   
   console.log(translatedData.todaysDate);
   console.log(`${project.name}: ${translatedData.totalHoursTracked} hrs`);
