+++ src/js/getTimeEntries.js
@@ -1,8 +1,5 @@
 const axios = require('axios');
-
-const HARVEST_TOKEN = '2719680.pt.8S1Yry48sFx_5qPZ4_a650DfG2Z4-_mwykp-uQTpWTVDttVS0CWxYah1wM_-mOdLNOdUpJ9QbdZFmap1UjgI_Q'; 
-const HARVEST_ACCOUNT_ID = 266954;
-const HARVEST_USER_ID = 3855900
+const { HARVEST_ACCOUNT_ID, HARVEST_TOKEN } = require('../constants/harvestInfo');
 
 const getTimeEntries = async ({ projectId }) => {
   const { data : { time_entries }} = await axios.get(`https://api.harvestapp.com/v2/time_entries?project_id=${projectId}&per_page=1`, {
